reworking the Colors class to build unique Logging color schemes
This ensures a unique Logging color scheme will be built per
`to_scheme` call.  No state is tracked on the Colors obj, just the
build to build the color scheme with.  The unique name of the color
scheme is returned from `to_scheme` so that the color scheme can be
referenced and used in layouts.