add `file_path` to logger API
This method returns the path for any file appenders that have been
added to the logger.  If no file appenders have been added or the
logger file appender api changes, this method just returns `nil`.

The goal here is to easily grab any log file path if you are logging
to a file with a logger.