remove ns-options as a dependency
This is part of getting this gem working in modern ruby versions.
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

NsOptions was used in a variety of contexts in this gem:

* I switched the main Settings to be a simple class that gets
  set to `nil` on reset
* I removed the NsOptions::Struct in the main intialize method in
  favor of just manually processing the opts
* I updated the both the colors and outputs to just manually
  define their data DSL meths and manually handle any defaults

This is prep or getting this gem working in modern ruby versions.