manually copy in Logging @ logging-1.8.2; add as `Logging182` module
This removes the logging v1.8.2 (the gem) as a dependency and
switches to manually copying the source from this version as
a `Logging182` submodule.  This is b/c the later versions are
not ruby 1.8.7 compatible.  However, version 1.8.2 is not ruby
1.9+ compatible.  I need Logsly to be both 1.8.7 *and* 1.9+
compatible, therefore I need to get a copy of the version 1.8.2
source and modify it to work in ruby 1.9+.

This copies the source as the submodule and tweaks the source
to allow requiring as `logsly/logging182`.  This involved updating
how the files are required internally and reworking some absolute
file path references.  As-is, this gets the test suite passing in
ruby 1.8.7 which is where things were when bring in Logging v1.8.2
as a gem dependency.

In a coming effort, I'll modify things to get the test suite
passing in modern ruby versions.