get implementation, tests working in modern Ruby versions
These are the things I needed to update to get the test suite
passing in modern Ruby versions.  I'll try and outline each change
below:

* force the `:outputs` option value to always be an array (we were
  getting away with passing a single string as `:outputs` b/c in
  ruby 1.8.7 String responds to `each` but not in modern Ruby)
* replace `Thread.exclusive` with a manual Mutex in Logging182's
  diagnostic context handling (`Thread.exclusive` is a thing in
  ruby 1.8.7 but not in modern ruby).
* fix a test in `output_tests.rb` to not blindly pass an the
  string representation of an `args` Array as a pattern (the test
  was getting away with this in ruby 1.8.7 b/c the string rep of
  an Array is just the string rep of the arrays items joined - in
  modern ruby the string rep of an Array looks more like the code
  rep, including brackets, etc - this broke the pattern handling
  for that test)
* make the base output tests more robust to catch some `Logging`
  renames I originally missed.

Note: I copied the Mutex handling from the latest version of
Logging to keep our copy as close to the original as possible.

Note also: In debugging that last pattern issue, I ended up
overwriting some of the pattern handling implementation with the
implementation from the latest version of Logging.  I feel this
implementation is much easier to read/understand and better
organizes the logic.