(hotfix) cleanup mdc and ndc use of Logging
This should have been done when I imported Logging and renamed it
to Logging182 back in PR 7.  This cleans up the missing renames
and adds tests to ensure we test this logic and prevent this type
of bug in the future.

This is all part of the larger effort to get logsly working in
modern ruby versions.