switch to creating a Logging182 Logger per output
This allows users to specify logger settings both on a global
and on a per-output basis.  For example, this allows saying one
output should log at the info level while another output logs at
the debug level.  Specifically, I want just this: Dk wants to use
a logger that logs to stdout on the INFO level while logging to
a file on the DEBUG level.

This switches to creating a Logging182 Logger per given output
and delegating all level method calls to each output's logger.
Note: this does potentially add a performance degredation since
each log statement has to run through multiple loggers under the
covers.

This also updates how outputs handle their data.  This is needed
b/c the data needs to be available on its own b/c it is needed
when building the output loggers.  Now the data is built and then
passed in to build the outputs appender. This adds a `level` DSL
method to the output datas so that it can be configured on
outputs. If an output's level is set, it will be used on the
output's logger. Otherwise, the logger will fall back to the
overall default level.

This change is backwards-compatible.  You can keep configuring
outputs as you did before and just setting the level on the logger
and nothing changes.  Everything works as it did before.  However,
you can now start specifying a level on your outputs and it will
be honored.

This also included a number of smaller cleanups:

* updated the README to show how output levels can be configured
* made the default level a formal constant
* removed the method missing delegation in favor of explicit
  Logging severity constant method definitions
* made the default output pattern a formal constant